Fix test for #4075 on 32-bit platforms

+++ test/core.jl
@@ -911,6 +911,6 @@ function foo4075(f::Foo4075, s::Symbol)
     x
 end
 
-@test isa(foo4075(Foo4075(1,2.0),:y), Float64)
+@test isa(foo4075(Foo4075(int64(1),2.0),:y), Float64)
 # very likely to segfault the second time if this is broken
-@test isa(foo4075(Foo4075(1,2.0),:y), Float64)
+@test isa(foo4075(Foo4075(int64(1),2.0),:y), Float64)
